+++ app/models/collection_object/dwc_extensions.rb
@@ -28,6 +28,9 @@ module CollectionObject::DwcExtensions
       verbatimEventDate: :dwc_verbatim_event_date,
       verbatimLocality: :dwc_verbatim_locality,
       waterBody: :dwc_water_body,
+      minimumDepthInMeters: :dwc_minimum_depth_in_meters,
+      maximumDepthInMeters: :dwc_maximum_depth_in_meters,
+      verbatimDepth: :dwc_verbatim_depth,
       identifiedBy: :dwc_identified_by,
       identifiedByID: :dwc_identified_by_id,
       dateIdentified: :dwc_date_identified,
@@ -201,12 +204,40 @@ module CollectionObject::DwcExtensions
     a.collect{|d| ApplicationController.helpers.label_for_taxon_determination(d)}.join(CollectionObject::DWC_DELIMITER).presence
   end
 
+  def dwc_internal_attribute_for(target = :collection_object, dwc_term_name)
+    return nil if dwc_term_name.nil?
+
+    case target
+    when  :collecting_event
+      return nil unless collecting_event
+      collecting_event.internal_attributes.includes(:predicate)
+        .where(
+          controlled_vocabulary_terms: {uri: ::DWC_ATTRIBUTE_URIS[dwc_term_name.to_sym] })
+        .pluck(:value)&.join(', ').presence
+    when :collection_object
+      internal_attributes.includes(:predicate)
+        .where(
+          controlled_vocabulary_terms: {uri: ::DWC_ATTRIBUTE_URIS[dwc_term_name.to_sym] })
+        .pluck(:value)&.join(', ').presence
+    else
+      nil
+    end
+  end
+
   def dwc_water_body
-    return nil unless collecting_event
-    collecting_event.internal_attributes.includes(:predicate)
-      .where(
-        controlled_vocabulary_terms: {uri: ::DWC_ATTRIBUTE_URIS[:waterBody] })
-      .pluck(:value)&.join(', ').presence
+    dwc_internal_attribute_for(:collecting_event, :waterBody)
+  end
+
+  def dwc_minimum_depth_in_meters
+    dwc_internal_attribute_for(:collecting_event, :minimumDepthInMeters)
+  end
+
+  def dwc_maximum_depth_in_meters
+    dwc_internal_attribute_for(:collecting_event, :maximumDepthInMeters)
+  end
+
+  def dwc_verbatim_depth
+    dwc_internal_attribute_for(:collecting_event, :verbatimDepth)
   end
 
   # TODO: consider CVT attributes with Predicates linked to URIs